examples/sar_workflow/requirements.txt
askdiana[app]>=0.4.0
python-dotenv>=0.19.0
requests>=2.20.0
